fix qt4.6 compilation
git-svn-id: svn://svn.code.sf.net/p/texstudio/code/trunk@2519 92cce386-e96d-4106-809f-596c673a3fd8

+++ grammarcheck.h
@@ -101,7 +101,7 @@ private:
 	
 	QString ltPath, javaPath;
 	QSet<QString> ignoredRules;
-	qint64 startTime;
+	uint startTime;
 	void tryToStart();
 	
 };
